移除 CMakeLists.txt 中的 C 标准设置；修复浮点数比较中的函数调用，使用 std::fabs 替代 std::fabsl。

+++ floating_point.cpp
@@ -22,7 +22,7 @@ std::ostream& operator<<(std::ostream& os, const float_& f) {
 }
 
 bool operator==(const float_& lhs, const float_& rhs) {
-  return std::fabsl(lhs.value() - rhs.value()) < float_::epsilon().value();
+  return std::fabs(lhs.value() - rhs.value()) < float_::epsilon().value();
 }
 
 bool operator<(const float_& lhs, const float_& rhs) {
